refactor: 패키지 참조 수정
